feature/updates-cli: feat: add unit tests for VTEX login, utility functions, and command execution
- Created tests for the login functionality, ensuring credentials are saved and errors are handled correctly.
- Added tests for the serviceGetAuth utility, covering various scenarios including successful and failed requests.
- Implemented tests for the run command utilities, verifying command execution and error handling.
- Introduced tests for shared utility functions such as account name extraction, debounce functionality, and OCLIF flag configurations.
- Added tests for handling multiple command executions and ensuring proper logging and error management.
- Created tests for version file utilities, validating version reading, updating, and script execution.
- Updated vitest configuration to include test files from the 'test' directory and adjusted coverage reporting settings.
