+++ .husky/commit-msg
@@ -1,3 +1,6 @@
-
 npx --no -- commitlint --edit ${1}
 
+branch="$(git rev-parse --abbrev-ref HEAD | sed 's/^origin\///')" 
+commitMsgFile=$1
+msg=$(cat "$commitMsgFile")
+echo "$branch: $msg" > "$commitMsgFile"
