--- src/modules/apps/pull-request/labels/suggest/suggest.test.ts
@@ -1,50 +1,0 @@
-import { suggestLabels } from '../index';
-import { SuggestLabelsOptions } from './types';
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-
-describe('suggestLabels', () => {
-  const mockOptions: SuggestLabelsOptions = {
-    branch: 'feature/test-branch',
-    target: 'main',
-    format: 'json',
-    verbose: false,
-  };
-
-  const mockOptionsTxt: SuggestLabelsOptions = {
-    branch: 'feature/test-branch',
-    target: 'main',
-    format: 'txt',
-    verbose: false,
-  };
-
-  beforeEach(() => {
-    // Mock console.log to avoid output during tests
-    vi.spyOn(console, 'log').mockImplementation(() => {});
-    vi.spyOn(console, 'error').mockImplementation(() => {});
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it('should be defined', () => {
-    expect(suggestLabels).toBeDefined();
-  });
-
-  it('should accept valid options', () => {
-    expect(() => {
-      // This should not throw
-      const options: SuggestLabelsOptions = {
-        target: 'main',
-        format: 'json',
-        verbose: true,
-      };
-      
-      const optionsTxt: SuggestLabelsOptions = {
-        target: 'main',
-        format: 'txt',
-        verbose: false,
-      };
-    }).not.toThrow();
-  });
-});
